+++ src/com/senla/carservice/controller/PlaceController.java
@@ -6,7 +6,7 @@ import com.senla.carservice.service.OrderService;
 import com.senla.carservice.service.OrderServiceImpl;
 import com.senla.carservice.service.PlaceService;
 import com.senla.carservice.service.PlaceServiceImpl;
-import com.senla.carservice.string.StringPlaces;
+import com.senla.carservice.stringutil.StringPlaces;
 import com.senla.carservice.util.DateUtil;
 
 import java.util.Date;
@@ -41,20 +41,22 @@ public class PlaceController {
     public String getArrayPlace() {
         try {
             return StringPlaces.getStringFromPlaces(placeService.getPlaces());
-        } catch (NumberObjectZeroException e) {
+        } catch (BusinessException e) {
             return e.getMessage();
         }
     }
 
     public String deletePlace(int index) {
         try {
-            placeService.deletePlace(placeService.getPlaces().get(index));
-            return String.format(" -delete place in service number \"%s\"",
-                    placeService.getPlaces().get(index).getNumber());
-        } catch (NumberObjectZeroException | BusinessException e) {
+            if (placeService.getPlaces().size() < index || index < 0 ) {
+                return "There are no such place";
+            } else {
+                placeService.deletePlace(placeService.getPlaces().get(index));
+                return String.format(" -delete place in service number \"%s\"",
+                        placeService.getPlaces().get(index).getNumber());
+            }
+        } catch ( BusinessException e) {
             return e.getMessage();
-        } catch (IndexOutOfBoundsException e){
-            return "There are no such place";
         }
     }
 
@@ -64,7 +66,7 @@ public class PlaceController {
         try {
             List<Order> orders = orderService.getOrderByPeriod(executeDate, leadDate);
             return StringPlaces.getStringFromPlaces(placeService.getFreePlaceByDate(executeDate, leadDate, orders));
-        } catch (NumberObjectZeroException | DateException | NullDateException e) {
+        } catch (BusinessException | DateException e) {
             return e.getMessage();
         }
     }
@@ -73,15 +75,16 @@ public class PlaceController {
         try {
             placeService.exportPlaces();
             return "Places have been export successfully!";
-        } catch (NumberObjectZeroException | ExportException e){
+        } catch (BusinessException e){
             return e.getMessage();
         }
     }
 
     public String importPlaces() {
         try {
-            return placeService.importPlaces();
-        } catch (NumberObjectZeroException e){
+            placeService.importPlaces();
+            return "places imported successfully";
+        } catch (BusinessException e){
             return e.getMessage();
         }
     }
@@ -90,7 +93,7 @@ public class PlaceController {
         try {
             placeService.serializePlace();
             return "Places have been serialize successfully!";
-        } catch (SerializeException e){
+        } catch (BusinessException e){
             return e.getMessage();
         }
     }
@@ -99,7 +102,7 @@ public class PlaceController {
         try {
             placeService.deserializePlace();
             return "Masters have been deserialize successfully!";
-        } catch (SerializeException e){
+        } catch (BusinessException e){
             return e.getMessage();
         }
     }
